requirements.txt
streamlit
fastapi
uvicorn
openai
python-dotenv
supabase
pydantic
python-multipart